Return string to display Origin depending on source of the assessment

Return string to display Origin depending on source of the assessment

Usage

get_pkg_origin(source)

Arguments

source

return value of function pkg_ref

Value

depending on source transformed string for the origin

Details

Please check riskmetric::pkg_ref_class_hierarchy and riskmetric::pkg_ref code to see the allowed values and structure